feat(annotate): wire AlphaMissense + ClinVar Parquet stores into the engine (3-way join)
Serve AlphaMissense and ClinVar from the per-source Parquet stores via chr-pruned DuckDB
joins, mirroring gnomad_parquet — giving ClinVar its first batch path (was ~24k per-variant
tabix + throttled NCBI):
- annotate/alphamissense_parquet.py + clinvar_parquet.py: prime() = one chr-pruned LEFT JOIN.
- alphamissense.prime() prefers the Parquet (tabix fallback preserved); clinvar.lookup() reads
  the Parquet cache before the tabix (a miss falls through unchanged — never a fake 'no record').
- pipeline primes ClinVar over the post-QC set. AlphaMissense stored as DOUBLE for bit-exact scores.
- Validated BYTE-IDENTICAL on real NA12878 (2,395 candidates: gene/tier/criteria/AM/ClinVar all
  match the tabix path, 0 diffs); ~12% faster offline. Fix AlphaMissense license label to CC BY-NC-SA.
- +6 tests (264 total); tabix-path tests isolated from the now-present Parquet store.

+++ scripts/build_alphamissense_parquet.py
@@ -47,7 +47,7 @@ COPY (
                 auto_detect=false, quote='', escape='', strict_mode=false,
                 columns={{'chrom':'VARCHAR','pos':'INTEGER','ref':'VARCHAR','alt':'VARCHAR',
                           'genome':'VARCHAR','uniprot_id':'VARCHAR','transcript_id':'VARCHAR',
-                          'protein_variant':'VARCHAR','am_pathogenicity':'FLOAT','am_class':'VARCHAR'}})
+                          'protein_variant':'VARCHAR','am_pathogenicity':'DOUBLE','am_class':'VARCHAR'}})
   GROUP BY chrom, pos, ref, alt
 ) TO '{TMP.as_posix()}' (FORMAT PARQUET, PARTITION_BY (chrom), OVERWRITE_OR_IGNORE);
 """)
